backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jqcc-backend"
version = "0.1.0"
description = "Jira Quality Command Center backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "gunicorn>=21.2",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "pydantic>=2.6",
  "pydantic-settings>=2.1",
  "python-jose[cryptography]>=3.3",
  "bcrypt>=4.0",
  "httpx>=0.26",
  "apscheduler>=3.10",
  "alembic>=1.13",
  "python-multipart>=0.0.9",
  "tenacity>=8.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.23",
  "respx>=0.21",
  "ruff>=0.4",
  "black>=24.1",
  "mypy>=1.8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["B008", "B904", "E501"]

[tool.black]
line-length = 100
target-version = ["py311"]
